auto_mobile/sentiment/requirements.txt
Appium-Python-Client==0.40
asn1crypto==0.24.0
blinker==1.4
brotlipy==0.7.0
certifi==2019.3.9
cffi==1.12.2
click==6.7
cryptography==2.3.1
h11==0.7.0
h2==3.1.0
hpack==3.0.0
hyperframe==5.2.0
idna==2.8
kaitaistruct==0.8
ldap3==2.5.2
mitmproxy==4.0.4
passlib==1.7.1
pika==1.0.0
pyasn1==0.4.5
pycparser==2.19
pyOpenSSL==18.0.0
pyparsing==2.2.2
pyperclip==1.6.5
PyYAML==5.1
redis==3.2.1
ruamel.yaml==0.15.91
selenium==3.141.0
six==1.12.0
sortedcontainers==2.0.5
tornado==5.1.1
urllib3==1.24.1
urwid==2.0.1
wsproto==0.11.0
